requirements.txt
python-telegram-bot==21.6
google-genai>=1.0.0
sqlalchemy==2.0.36
pillow==10.4.0
matplotlib==3.9.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.10.10
